Move db manager to core, stub out config manager
