+++ RealmListManager.UI/Core/ConfigurationManager.cs
@@ -1,0 +1,19 @@
+﻿namespace RealmListManager.UI.Core
+{
+    public class ConfigurationManager : IConfigurationManager
+    {
+        public bool RestoreRealmlist { get; set; }
+        public bool MinimizeToTray { get; set; }
+        public bool CloseToTray { get; set; }
+
+        public void Save()
+        {
+
+        }
+
+        public void Load()
+        {
+
+        }
+    }
+}
