+++ RealmListManager.UI/Core/IConfigurationManager.cs
@@ -1,0 +1,6 @@
+﻿namespace RealmListManager.UI.Core
+{
+    public interface IConfigurationManager
+    {
+    }
+}
