+++ RealmListManager.UI/RealmListManager.UI.csproj
@@ -101,13 +101,15 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="Core\ConfigurationManager.cs" />
     <Compile Include="Core\Converters\NegatedBooleanConverter.cs" />
     <Compile Include="Core\Entities\Location.cs" />
     <Compile Include="Core\Entities\Realmlist.cs" />
     <Compile Include="Core\Events\LocationChanged.cs" />
+    <Compile Include="Core\IConfigurationManager.cs" />
     <Compile Include="Core\IWindowConductor.cs" />
     <Compile Include="Core\ListBoxItemHelper.cs" />
-    <Compile Include="Core\Utilities\DbConnectionManager.cs" />
+    <Compile Include="Core\DbConnectionManager.cs" />
     <Compile Include="Core\Utilities\DbConnectionProvider.cs" />
     <Compile Include="Core\Utilities\ImageUtilities.cs" />
     <Compile Include="Core\Utilities\FileUtilities.cs" />
